+++ apps/admin-dashboard/src/app/layout.tsx
@@ -20,6 +20,10 @@ export const metadata: Metadata = {
     "Intelligent Traffic Management System - Administrative Control Panel",
 };
 
+import Sidebar from "@/components/admin/Sidebar";
+import TopBar from "@/components/admin/TopBar";
+import AlertBanner from "@/components/admin/AlertBanner";
+
 export default function RootLayout({
   children,
 }: Readonly<{ children: React.ReactNode }>) {
@@ -35,7 +39,24 @@ export default function RootLayout({
         />
       </head>
       <body className="bg-background text-on-background font-body-md text-body-md min-h-screen">
-        {children}
+        <div className="flex min-h-screen">
+          {/* Sidebar (fixed left, w-64) */}
+          <Sidebar />
+
+          {/* Main Content Area */}
+          <div className="ml-64 flex flex-col min-h-screen w-full">
+            {/* TopBar (sticky top) */}
+            <TopBar />
+
+            {/* AlertBanner (below topbar) */}
+            <AlertBanner />
+
+            {/* Page content area (scrollable) */}
+            <main className="flex-1 p-margin flex flex-col gap-margin overflow-y-auto">
+              {children}
+            </main>
+          </div>
+        </div>
       </body>
     </html>
   );
